Move css files to top level
