+++ src/stories/index.jsx
@@ -1,4 +1,4 @@
 import '@fontsource/raleway';
-import '../gen3-ui-component/css/base.css';
-import '../gen3-ui-component/css/icon.css';
+import '../base.css';
+import '../icon.css';
 import 'graphiql/graphiql.css';
